+++ apps/desktop/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pipelab/app
 
+## 2.0.0-beta.7
+
+### Patch Changes
+
+- fixes
+
 ## 2.0.0-beta.6
 
 ### Patch Changes