+++ assets/asset-discord/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pipelab/asset-discord
 
+## 1.0.0-beta.6
+
+### Patch Changes
+
+- fixes
+
 ## 1.0.0-beta.5
 
 ### Patch Changes