+++ plugins/plugin-discord/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pipelab/plugin-discord
 
+## 1.0.0-beta.10
+
+### Patch Changes
+
+- fixes
+- Updated dependencies
+  - @pipelab/plugin-core@1.0.0-beta.10
+  - @pipelab/shared@1.0.0-beta.6
+
 ## 1.0.0-beta.9
 
 ### Patch Changes